backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
geopy==2.4.1
shapely==2.0.2
requests==2.31.0
python-multipart==0.0.6
geojson==3.1.0


